fix broken command

+++ app/Console/Commands/FenixResolveCarPartImagesCommand.php
@@ -249,19 +249,20 @@ class FenixResolveCarPartImagesCommand extends Command
                     //$image->priority = $carImageNumber;
                     $image->save();
 
-                    if (file_exists($tempImagePath)) {
-                        unlink($tempImagePath);
-                    }
-
-                    if (file_exists($tempFilePath)) {
-                        unlink($tempFilePath);
-                    }
-//
-//
 
                 } catch (Exception $e) {
                     $this->error('Directory creation failed: ' . $e->getMessage());
                     return Command::FAILURE;
+                } finally {
+                    if (isset($tempImagePath) && file_exists($tempImagePath)) {
+                        unlink($tempImagePath);
+                    }
+
+                    if (isset($tempFilePath) && file_exists($tempFilePath)) {
+                        unlink($tempFilePath);
+                    }
+
+
                 }
